Check Description
Added a cache system so we know when a user last scanned their card. If the last scan was within 60 seconds or the same card was scanned twice outside of the time limit, the card will not register. Added an error sound that plays whenever something goes wrong with the scanner. Revamped the isSignIn function to work offline